requirements.txt

streamlit
pandas
openpyxl
fpdf2
arabic-reshaper
python-bidi
qrcode
